+++ src/plonemeeting/portal/core/browser/publication.py
@@ -1,3 +1,5 @@
+import datetime
+
 from imio.helpers.workflow import get_state_infos
 from imio.pyutils.utils import sort_by_indexes
 from plone import api
@@ -49,13 +51,13 @@ class PublicationView(DefaultView):
     """
 
     def __call__(self):
-        if api.content.get_state(self.context) == 'private' and \
-           _checkPermission(ModifyPortalContent, self.context) and \
-           self.context.enable_timestamping is False:
-            api.portal.show_message(
-                _("Timestamping is disabled for this element!"),
-                request=self.request,
-                type="warning")
+        if api.content.get_state(self.context) == "private" and _checkPermission(ModifyPortalContent, self.context):
+            if self.context.enable_timestamping is False:
+                api.portal.show_message(
+_("Timestamping is disabled for this element!"), request=self.request, type="warning"
+                )
+            if self.context.effective_date and self.context.effective_date.isPast():
+                api.portal.show_message(_("effective_date_in_past_msg"), request=self.request, type="warning")
         return super(PublicationView, self).__call__()
 
     def get_effective_date(self):